pyproject.toml
[project]
name = "scamper-optimizer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.4",
]
